legged_gym/sim_soccer2/simulation/isaac_sim/requirements.txt
# Simulation dependencies
flask
flask_socketio
numpy
gym
Pillow
setuptools
cli_args

# Remote communication
pyzmq